requirements.txt
opencv-python
mediapipe
numpy
scipy
fastapi
uvicorn
pydantic
sounddevice
pytest
paho-mqtt
pyyaml
